+++ packages/typedoc-plugin-markdown/src/theme/resources/partials/index.reflection.ts
@@ -31,6 +31,14 @@ export function reflectionIndex(
         });
       } else {
         md.push(heading(subHeadingLevel, reflectionGroup.title) + '\n');
+
+        /** CSDK START */
+        // Include groupDescription
+        if (reflectionGroup.description) {
+          md.push(reflectionGroup.description[0].text + '\n');
+        }
+        /** CSDK END */
+
         md.push(getGroup(context, reflectionGroup) + '\n');
       }
     });